Updated landlab for numpy 1.24, scipy 1.10 (#1521)
* use int instead of np.int in cython files

* ensure interp1d gets arrays with two elements

* fix ragged arrays for numpy 1.24

* remove lint

* add news fragment

+++ landlab/components/bedrock_landslider/cfuncs.pyx
@@ -2,7 +2,7 @@ cimport numpy as np
 cimport cython
 import numpy as np
 
-DTYPE_INT = np.int
+DTYPE_INT = int
 ctypedef np.int_t DTYPE_INT_t
 
 DTYPE_FLOAT = np.float64
